language menu is now responsive instead and adjusts heights accordingly
